apps/backend/pyproject.toml
[project]
name = "my-first-langgraph"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = "==3.12.*"
dependencies = [
    "descope>=1.7.6",
    "langchain[anthropic]>=0.3.25",
    "langgraph>=0.4.7",
    "langgraph-cli[inmem]>=0.2.10",
    "langgraph-api>=0.2.92",
    "python-dotenv>=1.1.0",
]

[dependency-groups]
dev = [
    "pyright>=1.1.403",
    "ruff>=0.12.5",
]
test = [
    "pytest>=8.3.4",
    "pytest-asyncio>=0.24.0",
    "respx>=0.21.1",
    "httpx>=0.28.1",
    "langchain-openai>=0.3.0",
]
